Isolating EventHubs authentication into a separate package (#504)
