+++ src/main/scala/com/databricks/labs/overwatch/pipeline/BronzeTransforms.scala
@@ -3,6 +3,7 @@ package com.databricks.labs.overwatch.pipeline
 import com.databricks.dbutils_v1.DBUtilsHolder.dbutils
 import com.databricks.labs.overwatch.BatchRunner.spark
 import com.databricks.labs.overwatch.env.Database
+import com.databricks.labs.overwatch.eventhubs.AadAuthInstance
 import com.databricks.labs.overwatch.utils.Helpers.getDatesGlob
 import com.databricks.labs.overwatch.utils.SchemaTools.structFromJson
 import com.databricks.labs.overwatch.utils._
@@ -211,6 +212,7 @@ trait BronzeTransforms extends SparkSessionWrapper {
                                     isFirstRun: Boolean,
                                     organizationId: String,
                                     runID: String): DataFrame = {
+    import com.databricks.labs.overwatch.eventhubs.AadClientAuthentication
 
     val connectionString = ConnectionStringBuilder(
       PipelineFunctions.parseAndValidateEHConnectionString(ehConfig.connectionString, ehConfig.azureClientId.isEmpty))
@@ -246,8 +248,7 @@ trait BronzeTransforms extends SparkSessionWrapper {
         "aad_client_id" -> PipelineFunctions.maybeGetSecret(ehConfig.azureClientId.get),
         "aad_client_secret" -> PipelineFunctions.maybeGetSecret(ehConfig.azureClientSecret.get),
         "aad_authority_endpoint" -> ehConfig.azureAuthEndpoint)
-      ehConf.setAadAuthCallbackParams(aadParams)
-        .setAadAuthCallback(new com.databricks.labs.overwatch.utils.AadClientAuthentication(aadParams))
+      AadAuthInstance.addAadAuthParams(ehConf, aadParams)
     } else
       ehConf
